requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
ultralytics
pillow
pandas
numpy
joblib
scikit-learn
python-multipart
